requirements.txt
ccxt
SQLAlchemy
websocket-client
python-dotenv
psycopg2